# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API Backend
  version: 1.0.0
  description: Production-ready Express + TypeScript API
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      summary: Health check endpoint
      responses:
        '200':
          description: Service health
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
  /api/auth/login:
    post:
      summary: Login user and return JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
  /api/auth/refresh:
    post:
      summary: Refresh access token
      responses:
        '200':
          description: Refreshed
  /api/auth/logout:
    post:
      summary: Logout user
      responses:
        '204':
          description: Logged out
  /api/auth/me:
    get:
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
  /api/users:
    post:
      summary: Create user (admin)
      security:
        - bearerAuth: []
      responses:
        '201':
          description: User created
    get:
      summary: List users with pagination
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: Paginated user list
  /api/users/{id}:
    get:
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User details
    put:
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated user
    delete:
      summary: Delete user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
  /api/users/{id}/profile:
    get:
      summary: Get user profile with extended info
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User profile with task stats
  /api/tasks:
    post:
      summary: Create task
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Task created
    get:
      summary: List user tasks with filtering/sorting
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: completed
          schema: { type: boolean }
        - in: query
          name: sortBy
          schema: { type: string }
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc] }
      responses:
        '200':
          description: Task list
  /api/tasks/{id}:
    get:
      summary: Get task by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Task details
    put:
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated task
    delete:
      summary: Delete task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
  /api/tasks/{id}/complete:
    patch:
      summary: Mark task as complete
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Completed task
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }